added check in meteo files instead of assuming same date
the function .set_value does not raise anything if it cannot find the
date in the file. So when timestep given as input by user different from
timestep in file, the script was running anyway.

To prevent that, the function get_dataframe_for_daily meteo now checks
for each timestep if value is in the file, if not the run is stopped