improvements in get_df_flow
the output file generated was including the first time_step of the time
series that is only here for initial conditions in simulation, so
removed it

also used paramater 'converters' in read_csv in order to convert the
flow data in string. That way, the except ValueError will be raised when
converting to float() when data is missing in input file (because
float('') raise it). Before the use of converter, the df contained NaN
for missing data, and float(NaN) does not raise a ValueError