added option to pour runoff up or down
The original decision I made was to pour the catchment runoff (i.e. the
flow reaching the river banks) into the node upstream of the link in
order to include a delaying component [i.e. overestimating reality]. But
for node0000 and node0001, it means that there is only a linres
reservoir linking the two and that node0000 has no catchment
contribution to its m3/s. It meant that the flow at node0001 was higher
than in node0000 because node0001 is input of the linear reservoir,
node0000 is the output

To be able to decide to pour into the downstream node, I added a
paramater to the Network __init__ to decide for upstream or downstream