renames step for gap where relevant
'time step' and 'time gap' were used interchangeably in the scripts
which could be a bit confusing at times.

harmonised this by changing the relevant mention of time step by time
gap.

the new harmonisation is based on this definition: 'time gap' in the
duration between two consecutive 'time steps'. A 'time step' is an exact
moment in time (precision of the second).

Simulations are carried out for each time step, which means that the
values of the variables can only be given for a precise time, however
the time gap preceding the time step has influenced the value of the
variables for the time step.