removes add_up parameter
this parameter became useless when I redesigned CSF to be compatible with SMARTpy (i.e. no delay in the runoff water being routed in the reservoir), maybe it was still possible to use add_down without any risk, but I did not test it, and adding down does not make much sense anyway, so I decided to get rid of it altogether.