redesigns communication with models to be more imperative-like
in preparation for the inclusion of the possibility to use SMARTc in CSF, needed to make the model run functions more imperative (so that a run function from C can be used in place of the Python function).

It meant getting rid of the communication between CSF and the models using dictionaries and asking the models to query in these to find the value it needs. Now the run functions get a tuple containing all the value it needs (inputs, parameters, and constants), and now returns also a tuple with all the processes, updated states, and outputs rather than storing them in the dataframe (nested dictionary).

New functions get_in/get_out in each model module needed to be created to be the communicator between the Python data structures an the in tuple and out tuple of the run functions of the model.

Also got rid of the direct communication between SMART catchment model and INCA on land that used to exist by passing a dict of dict, because it prevented the use SMARTc in place of the Python SMART implemented in CSF. But there was some information that was passed through these dict that could not be access from timestep and timestep-1 in the data structures, that is the effective rainfall contributing to the different pathways reservoirs during the timestep. Indeed, because the volume of each reservoir had the possibility to be updated if it goes negative, I could not do backwards calculations using old volume and new volume of reservoir to get the contribution of effective rainfall. This lead me to take the decision to include another type of variable in the Model object, termed "processes" (inspired from thermodynamics terminology, "process variable"). These processes are intended to store variables that represent fluxes happening during the timestep of a model run. This fluxes can then be used by other models that are applied on the same system (i.e. catchment) during the same timestep (typically the case here for the water quality model that needs info on this moving water during timestep to move contaminant during this same timestep). This meant defining 5 new processes (process variables) in the SMART model definition (docstring) and returning them in the data structure. To have room in these data structures to store the process variables, needed to update Model object, .spec file of each model to add a list "processes," whether it is empty of variables (INCAL, INCAS, LINRES) or not (SMART), and also modifying the generation of the data structures to use model.process_names alongside inputs, states, outputs.

Nota: The new implementation of run functions of the model takes mostly scalar variables in the argument tuple, but it also takes "waterbody" and "datetime" and "logger" in order to keep the reporting capabilities of the Python implementation of the model in the log file. This will not be the case if SMARTc or another C-extension is used. However, it is crucial for the get_in functions to return tuple containing strictly scalars, because get_in is intended to be used to send the input/states/parameters to both C-extension models and pure Python models.

Took the opportunity to add more docstrings everywhere to improved standalone documentation of CSF.