adds use of SMARTc if available
Adds the functionality to use SMARTc.onestep_c and SMARTc.onestep_r in place of smart and linres models available in CSF.

If import does not work (i.e. mostl likely due to it not being installed) it will fall back on using the pure python models available within CSF.

At the moment, only SMART is available as a C++ extension for Python, but because of the previous commits, it could be easy to link CSF to a C++ extension of INCA, if it were to exist in the future.

No proper benchmark done, but for a 2-year period on the Suir catchment (daily input data, hourly time step for simulation, 365 days of warm-up, slice-up 2000) the use of SMARTc is really only marginally faster (14m44s vs. 15m14s).  However, compared to the less imperative implementation of SMART LINRES and INCA that was originally developed, it seems to be faster (original version was taking 19m40s on the same catchment).