revamps TimeFrame + includes time rescale functions
1)
in order to allow for the reporting on a time gap different from simu or data time gap, needed to include more attibutes in TimeFrame for 'save'

this implied modifying the update_files functions to use this new save_timeseries

this also implied modifying the .simulation file to change some strings (replacing simu by save) and also adding another timegap for save, in addition to data_gap and simu_gap. So the setup_simulation function was modified accordingly. Took the occasion to remove all of the time consistency checks in this function, to move them in the TimeFrame class where it is more logical, and less hidden

2)
took this opportunity, working on temporal aspects, to include the rescale functions originating from SMARTpy. These functions are most versatile than what was avaialble in CSF (because an input data series were only processable if simu_start and data_start were the same, and if simu_gap <= data_gap). So now thanks to the use of gcd(gcd()), can determine the required resolution needed to rescale from data_gap to simu_gap given a potential time shift on the respective starts.

This implied modifying the read_csv and read_netcdf functions and inverting column/index in the data_nested_dict because it was in the format nd[datetime][variable]. It needed to be in the format nd[variable][datetime] because the rescale functions from SMARTpy needed to work on a single dict containing datetime for keys and data for values. However, it was required to keep the storage dataframe as nd_dataframe[datetime][variable] because communication between warm_up and simulation, and also between slices often use the update function of dict, so needed to have all variables for a given timestep in one single dict.